corrected some bugs, added correct lib

+++ src/ft_strlen.c
@@ -1,6 +1,8 @@
+#include "libft.h"
+
 size_t ft_strlen(const char *str, size_t maxlen)
 {
-    int i;
+    size_t i;
 
     i = 0;
     while (str[i] && i < maxlen)
